sol2.cfront: Added (int) cast to fix: 	semantics of "<" change in ANSI C; use explicit cast.

+++ ptclassic/src/octtools/Packages/options/options.c
@@ -124,7 +124,7 @@ char *argv[];
     if (scanPtr == NIL(char) || *scanPtr == '\0') {
 	if (		optind >= argc ||
 			argv[optind][0] != '-' ||
-			strlen(argv[optind]) < 2 ||
+			(int)strlen(argv[optind]) < 2 ||
 			strcmp(argv[optind], "--") == 0) {
 	    return(EOF);
 	}
